feat: refactor para utilizar mysql

+++ src/main/java/pos/fiap/lanchonete/adapter/out/database/PagamentoDbAdapter.java
@@ -1,14 +1,16 @@
-package pos.fiap.lanchonete.adapter.out.mongo;
+package pos.fiap.lanchonete.adapter.out.database;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import pos.fiap.lanchonete.adapter.out.mongo.entities.PagamentoEntity;
-import pos.fiap.lanchonete.adapter.out.mongo.entities.mapper.PagamentoEntityMapper;
-import pos.fiap.lanchonete.adapter.out.mongo.repository.PagamentoRepository;
+import pos.fiap.lanchonete.adapter.out.database.entities.PagamentoEntity;
+import pos.fiap.lanchonete.adapter.out.database.entities.mapper.PagamentoEntityMapper;
+import pos.fiap.lanchonete.adapter.out.database.repository.PagamentoRepository;
+import pos.fiap.lanchonete.adapter.out.database.repository.PedidoRepository;
 import pos.fiap.lanchonete.domain.model.DadosPagamento;
 import pos.fiap.lanchonete.port.PagamentoDbAdapterPort;
 
+import static java.lang.String.format;
 import static pos.fiap.lanchonete.utils.Constantes.FIM;
 import static pos.fiap.lanchonete.utils.Constantes.INICIO;
 
@@ -23,16 +25,24 @@ public class PagamentoDbAdapter implements PagamentoDbAdapterPort {
     private static final String STRING_LOG_FORMAT = "%s_%s_%s {}";
 
     private final PagamentoRepository pagamentoRepository;
+    private final PedidoRepository pedidoRepository;
     private final PagamentoEntityMapper pagamentoEntityMapper;
 
     @Override
     public DadosPagamento obterDadosPagamento(String idPedido) {
-        log.info(String.format(STRING_LOG_FORMAT, SERVICE_NAME, BUSCAR_STATUS_PAGAMENTO_METHOD_NAME, INICIO), idPedido);
+        log.info(format(STRING_LOG_FORMAT, SERVICE_NAME, BUSCAR_STATUS_PAGAMENTO_METHOD_NAME, INICIO), idPedido);
 
-        var pagamentoEntity = pagamentoRepository.findByIdPedido(idPedido);
+        var pedido = pedidoRepository.findById(idPedido);
+
+        if (pedido.isEmpty()) {
+            log.info(format(STRING_LOG_FORMAT, SERVICE_NAME, BUSCAR_STATUS_PAGAMENTO_METHOD_NAME, "_ERROR"), idPedido);
+            return null;
+        }
+
+        var pagamentoEntity = pagamentoRepository.findByPedidoEntity(pedido.get());
 
         if (pagamentoEntity.isPresent()) {
-            log.info(String.format(STRING_LOG_FORMAT, SERVICE_NAME, BUSCAR_STATUS_PAGAMENTO_METHOD_NAME, FIM), idPedido);
+            log.info(format(STRING_LOG_FORMAT, SERVICE_NAME, BUSCAR_STATUS_PAGAMENTO_METHOD_NAME, FIM), idPedido);
             return pagamentoEntityMapper.toDadosPagamento(pagamentoEntity.get());
         } else {
             return null;
@@ -41,14 +51,21 @@ public class PagamentoDbAdapter implements PagamentoDbAdapterPort {
 
     @Override
     public DadosPagamento salvarPagamento(DadosPagamento dadosPagamento) {
-        log.info(String.format(STRING_LOG_FORMAT, SERVICE_NAME, SALVAR_PAGAMENTO_METHOD_NAME, INICIO), dadosPagamento);
-        var pagamentoEntityFindById = pagamentoRepository.findByIdPedido(dadosPagamento.getDadosPedido().getNumeroPedido());
+        log.info(format(STRING_LOG_FORMAT, SERVICE_NAME, SALVAR_PAGAMENTO_METHOD_NAME, INICIO), dadosPagamento);
+        var pedido = pedidoRepository.findById(dadosPagamento.getDadosPedido().getNumeroPedido());
+
+        if (pedido.isEmpty()) {
+            log.info(format(STRING_LOG_FORMAT, SERVICE_NAME, SALVAR_PAGAMENTO_METHOD_NAME, "_ERROR"), dadosPagamento);
+            return null;
+        }
+
+        var pagamentoEntityFindById = pagamentoRepository.findByPedidoEntity(pedido.get());
         var pagamentoEntity = PagamentoEntity.builder().build();
 
         if (pagamentoEntityFindById.isPresent()) {
             // TODO Refatorar update
             pagamentoRepository.delete(pagamentoEntityFindById.get());
-            pagamentoEntityFindById.get().atualizarDadosEntity(dadosPagamento);
+            pagamentoEntityFindById.get().atualizarDadosEntity(dadosPagamento, pagamentoEntityFindById.get().getPedidoEntity());
             pagamentoEntity = pagamentoEntityFindById.get();
         } else {
             pagamentoEntity = pagamentoEntityMapper.toEntity(dadosPagamento);
@@ -56,7 +73,7 @@ public class PagamentoDbAdapter implements PagamentoDbAdapterPort {
 
         pagamentoEntity = pagamentoRepository.save(pagamentoEntity);
         var pagamento = pagamentoEntityMapper.toDadosPagamento(pagamentoEntity);
-        log.info(String.format(STRING_LOG_FORMAT, SERVICE_NAME, SALVAR_PAGAMENTO_METHOD_NAME, FIM), pagamento);
+        log.info(format(STRING_LOG_FORMAT, SERVICE_NAME, SALVAR_PAGAMENTO_METHOD_NAME, FIM), pagamento);
         return pagamento;
     }
 }
